更新2016_01_04_173148_create_admin_tables中integer为bigInteger

+++ database/migrations/2016_01_04_173148_create_admin_tables.php
@@ -49,8 +49,8 @@ class CreateAdminTables extends Migration
 
         Schema::create(config('admin.database.menu_table'), function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('parent_id')->default(0);
-            $table->integer('order')->default(0);
+            $table->bigInteger('parent_id')->default(0);
+            $table->bigInteger('order')->default(0);
             $table->string('title', 50);
             $table->string('icon', 50);
             $table->string('uri', 50)->nullable();
@@ -60,36 +60,36 @@ class CreateAdminTables extends Migration
         });
 
         Schema::create(config('admin.database.role_users_table'), function (Blueprint $table) {
-            $table->integer('role_id');
-            $table->integer('user_id');
+            $table->bigInteger('role_id');
+            $table->bigInteger('user_id');
             $table->index(['role_id', 'user_id']);
             $table->timestamps();
         });
 
         Schema::create(config('admin.database.role_permissions_table'), function (Blueprint $table) {
-            $table->integer('role_id');
-            $table->integer('permission_id');
+            $table->bigInteger('role_id');
+            $table->bigInteger('permission_id');
             $table->index(['role_id', 'permission_id']);
             $table->timestamps();
         });
 
         Schema::create(config('admin.database.user_permissions_table'), function (Blueprint $table) {
-            $table->integer('user_id');
-            $table->integer('permission_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('permission_id');
             $table->index(['user_id', 'permission_id']);
             $table->timestamps();
         });
 
         Schema::create(config('admin.database.role_menu_table'), function (Blueprint $table) {
-            $table->integer('role_id');
-            $table->integer('menu_id');
+            $table->bigInteger('role_id');
+            $table->bigInteger('menu_id');
             $table->index(['role_id', 'menu_id']);
             $table->timestamps();
         });
 
         Schema::create(config('admin.database.operation_log_table'), function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id');
+            $table->bigInteger('user_id');
             $table->string('path');
             $table->string('method', 10);
             $table->string('ip');
